E2e to cy/port/databases (#13018)
* done with 'table.cy.spec.js'

* done with 'edit.cy.spec.js'
