+++ frontend/test/metabase/scenarios/admin/databases/list.cy.spec.js
@@ -8,7 +8,18 @@ describe("scenarios > admin > databases > list", () => {
     cy.server();
   });
 
-  it("should let you view a database", () => {
+  it("should let you see databases in list view", () => {
+    cy.visit("/admin/databases");
+    cy.findByText("Sample Dataset");
+    cy.findByText("H2");
+  });
+
+  it("should not let you see saved questions in the database list", () => {
+    cy.visit("/admin/databases");
+    cy.get("td").should("have.length", 3);
+  });
+
+  it("should let you view a database's detail view", () => {
     cy.visit("/admin/databases");
     cy.contains("Sample Dataset").click();
     cy.url().should("match", /\/admin\/databases\/\d+$/);
@@ -18,9 +29,10 @@ describe("scenarios > admin > databases > list", () => {
     cy.visit("/admin/databases");
     cy.contains("Add database").click();
     cy.url().should("match", /\/admin\/databases\/create$/);
+    // *** code here should be more thorough
   });
 
-  it("should let you edit a database", () => {
+  it("should let you access edit page a database", () => {
     cy.visit("/admin/databases");
     cy.contains("Sample Dataset").click();
     cy.url().should("match", /\/admin\/databases\/1$/);
